Emit only sane pgid value for `jobs` output
We were previously printing the internal INVALID_PID value (since removed),
which was a meaningless `-2` constant, when there was no pgid associated with a
job.

This PR changes that to `-` to indicate no pgid available, which I prefer over
something like `0` or `-1`, but will cause problems for code that is hardcoded
to convert this field to an integral value.

Other options include 0 or -1 as mentioned above, or printing fish's own pgid.