Flush stdout when control returns to shell
This prevents kernel-buffered stdout contents from being written to the terminal
after the prompt itself, which can happen when the external process is
terminated with a signal and hasn't executed the libc/runtime-provided exit
handlers that flush stdout.

While it is possible for a libc-less application to exit normally without
flushing the stdout fd, the only way for "normal" applications to bypass the
atexit handlers would be to exit via abort() (i.e. SIGABRT) or in response to
some other signal; so we currently only flush stdout in response to such an
exit to reduce syscall overhead.