Make new ctrl-l logic more portable
The behavior of the new scrollback-push is not consistent across various
terminals and the functionality of ctrl-l is too important to break in the
default bindings. This hack should preserve the benefits of scrollback-push
while keeping the portability of the old clear-screen: first we ask the terminal
to scroll everything offscreen *then* we clear the screen. The addition of
clear-screen should be a no-op on any terminal that fully supports the desired
behavior of invoking the new scrollback-push.

(There are no real performance-related concerns with ctrl-l.)

This handles multiple possible issues (observed or otherwise) with scrollback-push:
* The screen contents are pushed back into scrollback but the screen isn't cleared
  (observed under Prompt/iOS, FreeBSD console session)
* termcap SF isn't supported
* termcap UP isn't supported ("Only a few terminal descriptions provide these
  commands, and most programs do not use them.")

It still breaks in an ugly way if terminal position reporting isn't supported,
but I'll address that separately.